backend/requirements.txt
fastapi
uvicorn
sqlalchemy
databases
pydantic
passlib
bcrypt
python-jose
asyncpg
psycopg2-binary
pytest
pymilvus
aiofiles
minio
sentence-transformers
langchain
langchain-community
pypdf
